+++ lib/hephaestus_ecto/migrations/postgres.ex
@@ -55,13 +55,13 @@ defmodule HephaestusEcto.Migrations.Postgres do
 
     repo = Map.get_lazy(opts, :repo, fn -> repo() end)
     escaped_prefix = Map.fetch!(opts, :escaped_prefix)
-    workflow_instances_table = qualified_table(escaped_prefix, "workflow_instances")
 
     query = """
     SELECT pg_catalog.obj_description(pg_class.oid, 'pg_class')
     FROM pg_class
     LEFT JOIN pg_namespace ON pg_namespace.oid = pg_class.relnamespace
-    WHERE quote_ident(pg_namespace.nspname) || '.' || quote_ident(pg_class.relname) = '#{workflow_instances_table}'
+    WHERE pg_class.relname = 'workflow_instances'
+    AND pg_namespace.nspname = '#{escaped_prefix}'
     """
 
     case repo.query(query, [], log: false) do
@@ -87,8 +87,8 @@ defmodule HephaestusEcto.Migrations.Postgres do
 
   defp record_version(_opts, 0), do: :ok
 
-  defp record_version(%{prefix: prefix}, version) do
-    execute("COMMENT ON TABLE #{qualified_table(prefix, "workflow_instances")} IS '#{version}'")
+  defp record_version(%{quoted_prefix: quoted}, version) do
+    execute("COMMENT ON TABLE #{quoted}.workflow_instances IS '#{version}'")
   end
 
   def qualified_table(prefix, table) do
@@ -101,6 +101,7 @@ defmodule HephaestusEcto.Migrations.Postgres do
     validate_prefix!(opts.prefix)
 
     opts
+    |> Map.put(:quoted_prefix, inspect(opts.prefix))
     |> Map.put(:escaped_prefix, String.replace(opts.prefix, "'", "\\'"))
   end